Node collapse detection bug in OSDI devices fixed. OpenVAF created models only write 'true' into the collapsed nodes map. Before setup this map must be cleared. It is cleared implicitly before first setup (instance data is zeroed at allocation). For all other setup invocations clearing must be manual.
Initial enumeration of nodes is moved from nodeOrdering() to
mapUnknowns().

setupWorker() of builtin models does not return information on
mapping and topology changes. This is done only by the setupWorker()
of a builtin instance.

Mutual inductance cannot rely on unknown numbers in setupWorker() to
check if its controlling nodes changed (unknowns are assigned after
setup). Therefore we rely on node IDs (names). The unknown numbers are
collected in bindCore() which is called after the unknowns are assigned
to nodes.